componentフォルダの中にdialogフォルダを作り、ac_single_option_dialogとac_yes_no_dialogを収納した
